backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.0
httpx==0.27.2

# Auth (OAuth2 password flow + JWT)
passlib[bcrypt]==1.7.4
bcrypt<4
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# MongoDB Atlas
motor==3.6.0
dnspython==2.6.1

# LangChain + Gemini (keep these loosely pinned to avoid source builds on newer Python)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0

# Document parsing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
